修改：角色管理级别逻辑完善

+++ src/main/java/cn/karelian/kas/services/interfaces/IRolesService.java
@@ -22,5 +22,5 @@ public interface IRolesService extends IKasService<Roles, Roles> {
 
 	public Result authorize(AuthorizeParam params) throws TransactionFailedException;
 
-	public Result delete(Integer id);
+	public Result delete(Byte id);
 }
